// apps/nativeapp/src/types/interface/slice.interface.ts
export interface SampleTree {
  project_id?: string
  site_id?: string
  tree_id: string
  species_guid: string
  intervention_id: string
  sloc_id: string
  count: number
  latitude: number
  longitude: number
  device_latitude: number
  device_longitude: number
  location_accuracy: string
  image_url: string
  cdn_image_url: string
  specie_name: string
  local_name: string
  specie_diameter: number
  specie_height: number
  tag_id: string
  plantation_date: number
  status_complete: boolean
  location_id: string
  tree_type: 'sample' | 'single'
  parent_id: string
  additional_details: string
  app_meta_data: string
  hid: string,
  history: History[],
  status: INTERVENTION_STATUS
  remeasurement_dates: RemeasurementDate
  remeasurement_requires: boolean,
  is_alive: boolean,
  fix_required: FIX_REQUIRED
  image_data: {
    latitude: number
    longitude: number
    imageUrl: string
    cdnImageUrl: string
    currentloclat: number
    currentloclong: number
    isImageUploaded: boolean
    coordinateID: string
  },
}

export interface PlantedSpecies {
  id?: string,
  guid: string,
  scientificName: string,
  aliases: string,
  count: number,
  image: string
}
// ...
export interface InterventionLocation {
  type: string
  coordinates: string
}

export interface RemeasurementDate {
  sampleTreeId: string
  created: number
  lastMeasurement: number
  remeasureBy: number
  nextMeasurement: number
}

export interface History {
  history_id: string
  eventName: string
  eventDate: number
  imageUrl: string
  cdnImageUrl: string
  diameter: number
  height: number
  additionalDetails: any
  appMetadata: string
  status: string
  statusReason: string
  dataStatus: 'SYNCED' |  "REMEASUREMENT_DATA_UPLOAD" | "REMEASUREMENT_EVENT_UPDATE" | "SKIP_REMEASUREMENT"
  parentId: string
  samplePlantLocationIndex: number
  lastScreen: string
}

export interface InterventionData {
  form_id: string,
  intervention_id: string
  intervention_key: INTERVENTION_TYPE
  intervention_title: string
  intervention_date: number
  intervention_end_date: number
  project_id: string
  project_name: string
  site_name: string
  location_type: string
  location: InterventionLocation
  image: string
  image_data: [],
  has_species: boolean
  planted_species: PlantedSpecies[]
  has_sample_trees: boolean
  sample_trees: SampleTree[]
  is_complete: boolean
  site_id: string
  intervention_type: INTERVENTION_TYPE
  form_data: FormElement[]
  additional_data: FormElement[]
  meta_data: string
  last_updated_at?: number
  status: INTERVENTION_STATUS,
  hid: string,
  coords: {
    type: 'Point',
    coordinates: number[]
  },
  entire_site: boolean
  active?: boolean
  last_screen: LAST_SCREEN
  location_id: string
  locate_tree: string
  remeasurement_required: boolean
  next_measurement_date: number
  fix_required: FIX_REQUIRED
  is_legacy?: boolean
}